requirements.txt
pytest>=7.4.0
pytest-asyncio>=0.23.0
pyyaml>=6.0
jsonschema>=4.19.0
requests>=2.31.0
